+++ packages/games/src/lib/solo/coin-flip/components/bet-controller/strategy-controller.tsx
@@ -26,6 +26,7 @@ interface StrategyControllerProps {
   maxWager: number;
   isAutoBetMode: boolean;
   strategy: StrategyProps;
+  isPinNotFound?: boolean;
   onAutoBetModeChange: React.Dispatch<React.SetStateAction<boolean>>;
   onLogin?: () => void;
 }
@@ -35,6 +36,7 @@ export const StrategyController = ({
   maxWager,
   isAutoBetMode,
   strategy,
+  isPinNotFound,
   onAutoBetModeChange,
   onLogin,
 }: StrategyControllerProps) => {
@@ -130,7 +132,7 @@ export const StrategyController = ({
         </Button>
       )}
 
-      <PreBetButton onLogin={onLogin} className="wr-mb-3 lg:wr-mb-0">
+      <PreBetButton onLogin={onLogin} isPinNotFound={isPinNotFound} className="wr-mb-3 lg:wr-mb-0">
         <Button
           type={!isAutoBetMode ? 'button' : 'submit'}
           variant={'success'}