+++ packages/games/src/lib/solo/dice/components/bet-controller/index.tsx
@@ -18,10 +18,11 @@ interface Props {
   winMultiplier: number;
   isGettingResults?: boolean;
   isAutoBetMode: boolean;
+  isPinNotFound?: boolean;
   onAutoBetModeChange: React.Dispatch<React.SetStateAction<boolean>>;
   onBetModeChange: React.Dispatch<React.SetStateAction<BetMode>>;
-  strategy: StrategyProps;
   onLogin?: () => void;
+  strategy: StrategyProps;
 }
 
 export const BetController: React.FC<Props> = (props) => {