+++ packages/games/src/lib/solo/slots/winr-of-olympus-1000/index.ts
@@ -1,0 +1,3 @@
+export * from './store';
+export { WinrOfOlympus1000Template as WinrOfOlympus1000Template } from './template';
+export * from './types';