+++ packages/games/src/lib/solo/slots/winr-of-olympus-1000/types/index.ts
@@ -1,0 +1,5 @@
+export interface WinrOfOlympus1000FormFields {
+  betAmount: number;
+  actualBetAmount: number;
+  isDoubleChance: boolean;
+}