+++ packages/web3-games/src/games/coinflip-2d/index.tsx
@@ -15,6 +15,7 @@ import {
   useFastOrVerified,
   usePriceFeed,
   useSendTx,
+  useSessionStore,
   useTokenAllowance,
   useTokenBalances,
   useTokenStore,
@@ -292,6 +293,10 @@ export default function CoinFlipGame(props: TemplateWithWeb3Props) {
     [coinFlipResult]
   );
 
+  const sessionStore = useSessionStore();
+  const isPinNotFound =
+    (!sessionStore.pin || !localStorage['session-store']) && !currentAccount.isSocialLogin;
+
   const onAutoBetModeChange = () => clearIterationIntervals();
 
   React.useEffect(() => {
@@ -320,6 +325,7 @@ export default function CoinFlipGame(props: TemplateWithWeb3Props) {
           updateBetCondition: handleUpdateBetCondition,
           updateProfitCondition: handleUpdateProfitCondition,
         }}
+        isPinNotFound={isPinNotFound}
       />
       {!props.hideBetHistory && (
         <BetHistoryTemplate