+++ packages/web3-games/src/games/keno/index.tsx
@@ -15,6 +15,7 @@ import {
   useFastOrVerified,
   usePriceFeed,
   useSendTx,
+  useSessionStore,
   useTokenAllowance,
   useTokenBalances,
   useTokenStore,
@@ -292,6 +293,10 @@ export default function KenoGame(props: TemplateWithWeb3Props) {
 
   const onAutoBetModeChange = () => clearIterationIntervals();
 
+  const sessionStore = useSessionStore();
+  const isPinNotFound =
+    (!sessionStore.pin || !localStorage['session-store']) && !currentAccount.isSocialLogin;
+
   React.useEffect(() => {
     return () => {
       clearLiveResults();
@@ -317,6 +322,7 @@ export default function KenoGame(props: TemplateWithWeb3Props) {
           updateBetCondition: handleUpdateBetCondition,
           updateProfitCondition: handleUpdateProfitCondition,
         }}
+        isPinNotFound={isPinNotFound}
       />
       {!props.hideBetHistory && (
         <BetHistoryTemplate