+++ packages/web3-games/src/games/plinko/index.tsx
@@ -16,6 +16,7 @@ import {
   useFastOrVerified,
   usePriceFeed,
   useSendTx,
+  useSessionStore,
   useTokenAllowance,
   useTokenBalances,
   useTokenStore,
@@ -296,6 +297,10 @@ export default function PlinkoGame(props: TemplateWithWeb3Props) {
 
   const onAutoBetModeChange = () => clearIterationIntervals();
 
+  const sessionStore = useSessionStore();
+  const isPinNotFound =
+    (!sessionStore.pin || !localStorage['session-store']) && !currentAccount.isSocialLogin;
+
   React.useEffect(() => {
     return () => {
       clearLiveResults();
@@ -321,6 +326,7 @@ export default function PlinkoGame(props: TemplateWithWeb3Props) {
           updateBetCondition: handleUpdateBetCondition,
           updateProfitCondition: handleUpdateProfitCondition,
         }}
+        isPinNotFound={isPinNotFound}
       />
       {!props.hideBetHistory && (
         <BetHistoryTemplate