+++ packages/web3-games/src/games/rps/index.tsx
@@ -11,11 +11,11 @@ import {
 } from '@winrlabs/games';
 import {
   controllerAbi,
-  ErrorCode,
   useCurrentAccount,
   useFastOrVerified,
   usePriceFeed,
   useSendTx,
+  useSessionStore,
   useTokenAllowance,
   useTokenBalances,
   useTokenStore,
@@ -297,6 +297,10 @@ export default function RpsGame(props: TemplateWithWeb3Props) {
 
   const onAutoBetModeChange = () => clearIterationIntervals();
 
+  const sessionStore = useSessionStore();
+  const isPinNotFound =
+    (!sessionStore.pin || !localStorage['session-store']) && !currentAccount.isSocialLogin;
+
   React.useEffect(() => {
     return () => {
       clearLiveResults();
@@ -323,6 +327,7 @@ export default function RpsGame(props: TemplateWithWeb3Props) {
           updateBetCondition: handleUpdateBetCondition,
           updateProfitCondition: handleUpdateProfitCondition,
         }}
+        isPinNotFound={isPinNotFound}
       />
       {!props.hideBetHistory && (
         <BetHistoryTemplate