+++ apps/web/CHANGELOG.md
@@ -1,5 +1,25 @@
 # web
 
+## 3.1.5
+
+### Patch Changes
+
+- Updated dependencies [895133e]
+  - @winrlabs/games@5.1.3
+  - @winrlabs/web3@0.11.85
+  - @winrlabs/web3-games@13.0.5
+  - @winrlabs/ui@10.0.85
+
+## 3.1.4
+
+### Patch Changes
+
+- Updated dependencies [0219653]
+  - @winrlabs/games@5.1.2
+  - @winrlabs/web3@0.11.84
+  - @winrlabs/web3-games@13.0.4
+  - @winrlabs/ui@10.0.84
+
 ## 3.1.3
 
 ### Patch Changes