+++ packages/games/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @winrlabs/games
 
+## 5.1.3
+
+### Patch Changes
+
+- 895133e: winr of olympus 1000
+
+## 5.1.2
+
+### Patch Changes
+
+- 0219653: create a session text
+
 ## 5.1.1
 
 ### Patch Changes