+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @winrlabs/web3
 
+## 0.11.85
+
+### Patch Changes
+
+- 895133e: winr of olympus 1000
+
+## 0.11.84
+
+### Patch Changes
+
+- 0219653: create a session text
+
 ## 0.11.83
 
 ### Patch Changes